+++ src/textual_textarea/messages.py
@@ -4,20 +4,6 @@ from typing import Union
 from textual.message import Message
 
 
-class TextAreaCursorMoved(Message, bubble=True):
-    """Posted when the cursor moves
-
-    Attributes:
-        cursor_x: The x position of the cursor
-        cursor_y: The y position (line number)
-    """
-
-    def __init__(self, cursor_x: int, cursor_y: int) -> None:
-        super().__init__()
-        self.cursor_x = cursor_x
-        self.cursor_y = cursor_y
-
-
 class TextAreaScrollOne(Message, bubble=True):
     """
     Posted to get parent container to scroll one in a direction
@@ -47,3 +33,7 @@ class TextAreaSaved(Message, bubble=True):
     def __init__(self, path: Union[Path, str]) -> None:
         self.path = str(path)
         super().__init__()
+
+
+class TextAreaHideCompletionList(Message):
+    pass